Extracted all frontend methods
