+++ templates/frontend.rb
@@ -1,5 +1,17 @@
-def setup_govuk_frontend
+def initialize_formbuilder
+  return if file_contains?("config/initializers/govuk_formbuilder.rb", "GOVUKDesignSystemFormBuilder")
 
+  inject_into_file(
+    "app/controllers/application_controller.rb",
+    "default_form_builder(GOVUKDesignSystemFormBuilder::FormBuilder)\n".indent(2),
+    after: "class ApplicationController < ActionController::Base\n"
+  )
+
+  template('config/initializers/govuk_formbuilder.rb')
+end
+
+def setup_govuk_frontend
+  apply "templates/frontend/govuk_frontend.rb"
 end
 
 def setup_govuk_components
@@ -9,6 +21,8 @@ end
 def setup_govuk_formbuilder
   gem "govuk_design_system_formbuilder" unless
     file_contains?("Gemfile", "govuk_design_system_formbuilder")
+
+  initialize_formbuilder
 end
 
 def apply_template!
